x64: sys/posix/posix_main.cpp: warning: cast from 'void*' to 'int' loses precision

+++ neo/sys/posix/posix_main.cpp
@@ -416,7 +416,7 @@ Sys_DLL_Load
 TODO: OSX - use the native API instead? NSModule
 =================
 */
-int Sys_DLL_Load(const char *path)
+intptr_t Sys_DLL_Load(const char *path)
 {
 	void *handle = dlopen(path, RTLD_NOW);
 
@@ -424,7 +424,7 @@ int Sys_DLL_Load(const char *path)
 		Sys_Printf("dlopen '%s' failed: %s\n", path, dlerror());
 	}
 
-	return (int)handle;
+	return (intptr_t)handle;
 }
 
 /*
@@ -432,7 +432,7 @@ int Sys_DLL_Load(const char *path)
 Sys_DLL_GetProcAddress
 =================
 */
-void *Sys_DLL_GetProcAddress(int handle, const char *sym)
+void *Sys_DLL_GetProcAddress(intptr_t handle, const char *sym)
 {
 	const char *error;
 	void *ret = dlsym((void *)handle, sym);
@@ -449,7 +449,7 @@ void *Sys_DLL_GetProcAddress(int handle, const char *sym)
 Sys_DLL_Unload
 =================
 */
-void Sys_DLL_Unload(int handle)
+void Sys_DLL_Unload(intptr_t handle)
 {
 	dlclose((void *)handle);
 }
