+++ neo/sys/sys_local.h
@@ -62,9 +62,9 @@ class idSysLocal : public idSys
 		virtual bool			LockMemory(void *ptr, int bytes);
 		virtual bool			UnlockMemory(void *ptr, int bytes);
 
-		virtual int				DLL_Load(const char *dllName);
-		virtual void 			*DLL_GetProcAddress(int dllHandle, const char *procName);
-		virtual void			DLL_Unload(int dllHandle);
+		virtual intptr_t		DLL_Load(const char *dllName);
+		virtual void 			*DLL_GetProcAddress(intptr_t dllHandle, const char *procName);
+		virtual void			DLL_Unload(intptr_t dllHandle);
 		virtual void			DLL_GetFileName(const char *baseName, char *dllName, int maxLength);
 
 		virtual sysEvent_t		GenerateMouseButtonEvent(int button, bool down);
