+++ neo/sys/sys_public.h
@@ -325,9 +325,9 @@ const char 	*Sys_GetCallStackCurAddressStr(int depth);
 void			Sys_ShutdownSymbols(void);
 
 // DLL loading, the path should be a fully qualified OS path to the DLL file to be loaded
-int				Sys_DLL_Load(const char *dllName);
-void 			*Sys_DLL_GetProcAddress(int dllHandle, const char *procName);
-void			Sys_DLL_Unload(int dllHandle);
+intptr_t		Sys_DLL_Load(const char *dllName);
+void 			*Sys_DLL_GetProcAddress(intptr_t dllHandle, const char *procName);
+void			Sys_DLL_Unload(intptr_t dllHandle);
 
 // event generation
 void			Sys_GenerateEvents(void);
@@ -564,9 +564,9 @@ class idSys
 		virtual const char 	*GetCallStackCurStr(int depth) = 0;
 		virtual void			ShutdownSymbols(void) = 0;
 
-		virtual int				DLL_Load(const char *dllName) = 0;
-		virtual void 			*DLL_GetProcAddress(int dllHandle, const char *procName) = 0;
-		virtual void			DLL_Unload(int dllHandle) = 0;
+		virtual intptr_t		DLL_Load(const char *dllName) = 0;
+		virtual void 			*DLL_GetProcAddress(intptr_t dllHandle, const char *procName) = 0;
+		virtual void			DLL_Unload(intptr_t dllHandle) = 0;
 		virtual void			DLL_GetFileName(const char *baseName, char *dllName, int maxLength) = 0;
 
 		virtual sysEvent_t		GenerateMouseButtonEvent(int button, bool down) = 0;
